--- src/models/fundingSource.ts
@@ -1,3 +1,0 @@
-export enum FundingSource {
-    CDC = 'CDC'
-}
