--- src/models/fundingTime.ts
@@ -1,5 +1,0 @@
-export enum FundingTime {
-    Full = 'Full',
-    Part = 'Part',
-    Split = 'Split'
-}
