--- src/models/fundingTimeSplit.ts
@@ -1,9 +1,0 @@
-import { FundingSpace } from './fundingSpace';
-
-export interface FundingTimeSplit { 
-    id: number;
-    fundingSpaceId: number;
-    fundingSpace?: FundingSpace;
-    fullTimeWeeks: number;
-    partTimeWeeks: number;
-}
