+++ src/models/index.ts
@@ -4,19 +4,11 @@ export * from './cdcReport';
 export * from './child';
 export * from './enrollment';
 export * from './family';
-export * from './familyDetermination';
 export * from './funding';
-export * from './fundingSource';
-export * from './fundingSpace';
-export * from './fundingTime';
-export * from './fundingTimeSplit';
-export * from './fundingTimeSplitUtilization';
 export * from './gender';
 export * from './organization';
-export * from './organizationPermission';
 export * from './problemDetails';
 export * from './region';
 export * from './reportingPeriod';
 export * from './site';
-export * from './sitePermission';
 export * from './user';
