--- src/models/organizationPermission.ts
@@ -1,10 +1,0 @@
-import { Organization } from './organization';
-import { User } from './user';
-
-export interface OrganizationPermission { 
-    organizationId: number;
-    organization?: Organization;
-    id: number;
-    userId: number;
-    user?: User;
-}
