--- src/models/sitePermission.ts
@@ -1,10 +1,0 @@
-import { Site } from './site';
-import { User } from './user';
-
-export interface SitePermission { 
-    siteId: number;
-    site?: Site;
-    id: number;
-    userId: number;
-    user?: User;
-}
